+++ packages/graph-explorer/src/connector/openCypher/templates/verticesSchemaTemplate.ts
@@ -1,0 +1,6 @@
+const verticesSchemaTemplate = ({ type }: { type: string }) => {
+
+  return `MATCH (v:\`${type}\`) RETURN v AS object LIMIT 1`;
+};
+
+export default verticesSchemaTemplate;